Pinball can't leave map anymore, and new power
Pinball used to be able to bounce out of the arena easily, so I made a glass roof for the aesthetic (though it wasn't fully capable of keeping the ball in). And I set the Y position as static, meaning it can't change, thus the ball can't bounce up out of the arena. Plus, I added walls so it can't bounce out either.

Lastly, I added a duplicate power up script so that we can get two balls now! Also added weight and drag so the pinball should behave as you'd expect irl.